Created different files for routings and provided the proper routes, moreover connection with database within the index.js file itself.
