+++ backend/.env.example
@@ -1,2 +1,3 @@
 MONGO_URL = ""  # Paste your MongoDB URL here for local testing
 JWT_SECRET = ""   #Paste your JWT Secret here
+PORT = 5000
