--- backend/db.js
@@ -1,16 +1,0 @@
-const mongoose = require('mongoose');
-
-const connectDB = async () => {
-  try {
-    const conn = await mongoose.connect(process.env.MONGO_URL, {
-      useNewUrlParser: true,  // Used the new URL string parser
-      useUnifiedTopology: true, // Used the new connection management engine
-    });
-    console.log(`Connected to MongoDB Database: ${conn.connection.host}`);
-  } catch (error) {
-    console.error(`Error connecting to MongoDB: ${error.message}`);
-    process.exit(1); 
-  }
-};
-
-module.exports = connectDB;
