+++ backend/models/productsModel.js
@@ -1,0 +1,10 @@
+const mongoose = require('mongoose');
+
+const productSchema = new mongoose.Schema({
+  name: { type: String, required: true },
+  price: { type: Number, required: true },
+  description: { type: String, required: true },
+  imageUrl: { type: String },
+});
+
+const Product = mongoose.model('Product', productSchema);
