+++ backend/package.json
@@ -7,6 +7,7 @@
    "dependencies": {
       "bcrypt": "^5.1.1",
       "bcryptjs": "^2.4.3",
+      "body-parser": "^1.20.3",
       "cookie-parser": "^1.4.6",
       "cors": "^2.8.5",
       "dotenv": "^16.4.5",
