+++ frontend/src/components/User/ForgotPassword.jsx
@@ -1,4 +1,3 @@
-"use client";
 import React, { useState } from "react";
 import {
   Box,
